Add files via upload
This is the base script, I would like to add a few additional features such as:

a. Loop to bring you back to the main part of the script
b. Verification if PowerCLI is installed already before it tries to run the script. 

This should cut down on the amount of time it takes to run the script